Complicate lambda workflow to better demonstrate partial transpiling
Make putItem function not to write data directly to the database, but place it to a new queue instead.

Add new writeItem function that takes item from the queue and writes it to the database.

Add two new synthetic utility classes (for DynamoDB and SQS) that are used by some (but not all) lambdas.

Now, it is much more demonstrative that only relevant code is placed in each Laambda function.

With current project structure:

	src
	├── handlers
	│   ├── get-all-items.ts
	│   ├── get-by-id.ts
	│   ├── put-item.ts
	│   └── write-item.ts
	└── utils
	    ├── dynamodb.ts
	    └── sqs.ts

If we build putItem and writeItem functions we will get totally different set of files inside them:

	.aws-sam/build/putItemFunction
	└── dist
	    ├── handlers
	    │   └── put-item.js
	    └── utils
	        └── sqs.js

	.aws-sam/build/writeItemFunction
	└── dist
	    ├── handlers
	    │   └── write-item.js
	    └── utils
	        └── dynamodb.js